refactor(build-notification): split GitHub and GitLab code
